[FIX] Windows build

+++ src/client/scene/lobby/Lobby.cpp
@@ -40,7 +40,7 @@ void LobbyScene::run(NetworkClient &client, sf::RenderWindow &window, SceneScree
 
 void LobbyScene::init_lobby(sf::RenderWindow &window)
 {
-    this->_world.create_text("CHOOSE YOUR SPACESHIP", "assets/font/EMINOR-BlackItalic.ttf", 70, Vector2f{window.getSize().x / 2 - 367.5, 200.0});
+    this->_world.create_text("CHOOSE YOUR SPACESHIP", "assets/font/EMINOR-BlackItalic.ttf", 70, Vector2f{(float)window.getSize().x / 2 - 367.5f, 200.0});
     this->_world.create_drawable_object("assets/background/lobby.png", Vector4i{0, 0, 1056, 672}, Vector2f{1.9, 1.6}, Vector2f{0, 0});
 
     this->_world.create_drawable_object("assets/SpaceShip/ship_armored_spritesheet.png", Vector4i{0, 0, 128, 128}, Vector2f{2.0, 2.0}, Vector2f{164, (float)window.getSize().y / 2});
